Moved MariaDbContainerBaseTest to ../features/base + fix imports
